Check for quit without breaking input

+++ open-win.cpp
@@ -11,7 +11,7 @@ namespace ufo
     void Ufo::openFile(const file& file)
     {
        auto retval = ShellExecute(nullptr, "open", file.name.data(), nullptr, nullptr, SW_SHOWNORMAL);
-       if ((long)retval <= 32)
+       if ((size_t)retval <= 32)
        {
            std::cout << "Error opening file" << std::endl;
        }
